@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"reflect"
@@ -67,13 +66,35 @@ func (c *Client) BuildRequestUrl(path string, query map[string]interface{}) *url
 	return &u
 }
 
+func (c *Client) BuildRequestUrlWithRawQuery(path string, rawQuery string) *url.URL {
+	u := *c.endpoint
+	u.Path += path
+	u.RawQuery = rawQuery
+	return &u
+}
+
 func (c *Client) Get(path string, query map[string]interface{}, header http.Header) ([]byte, error) {
 	u := c.BuildRequestUrl(path, query)
 	return c.sendHttp("GET", u.String(), header, nil)
 }
 
+// GetString
+//
+// 某些api对参数顺序有要求, 只能自行构建query字符串
+func (c *Client) GetString(path string, rawQuery string, header http.Header) ([]byte, error) {
+	u := c.BuildRequestUrlWithRawQuery(path, rawQuery)
+	return c.sendHttp("GET", u.String(), header, nil)
+}
+
 func (c *Client) PostForm(path string, params map[string]interface{}, header http.Header) ([]byte, error) {
 	return c.sendFormRequest("POST", path, params, header)
+}
+
+// PostFormString
+//
+// 某些api对参数顺序有要求, 只能自行构建body字符串
+func (c *Client) PostFormString(path string, body string, header http.Header) ([]byte, error) {
+	return c.sendFormString("POST", path, body, header)
 }
 
 func (c *Client) PostJson(path string, params map[string]interface{}, header http.Header) ([]byte, error) {
@@ -84,12 +105,26 @@ func (c *Client) PutForm(path string, params map[string]interface{}, header http
 	return c.sendFormRequest("PUT", path, params, header)
 }
 
+// PutFormString
+//
+// 某些api对参数顺序有要求, 只能自行构建body字符串
+func (c *Client) PutFormString(path string, body string, header http.Header) ([]byte, error) {
+	return c.sendFormString("PUT", path, body, header)
+}
+
 func (c *Client) PutJson(path string, params map[string]interface{}, header http.Header) ([]byte, error) {
 	return c.sendJsonRequest("PUT", path, params, header)
 }
 
 func (c *Client) DeleteForm(path string, params map[string]interface{}, header http.Header) ([]byte, error) {
 	return c.sendFormRequest("DELETE", path, params, header)
+}
+
+// DeleteFormString
+//
+// 某些api对参数顺序有要求, 只能自行构建body字符串
+func (c *Client) DeleteFormString(path string, body string, header http.Header) ([]byte, error) {
+	return c.sendFormString("DELETE", path, body, header)
 }
 
 func (c *Client) DeleteJson(path string, params map[string]interface{}, header http.Header) ([]byte, error) {
@@ -111,7 +146,7 @@ func (c *Client) sendHttp(verb, u string, header http.Header, body io.Reader) ([
 	}
 	defer resp.Body.Close()
 
-	respBody, _ := ioutil.ReadAll(resp.Body)
+	respBody, _ := io.ReadAll(resp.Body)
 	if resp.StatusCode != http.StatusOK {
 		return respBody, fmt.Errorf(resp.Status)
 	}
@@ -122,20 +157,7 @@ func (c *Client) sendHttp(verb, u string, header http.Header, body io.Reader) ([
 func (c *Client) sendFormRequest(verb, path string, params map[string]interface{}, header http.Header) ([]byte, error) {
 	values := MapToUrlValues(params)
 	data := values.Encode()
-	body := strings.NewReader(data)
-	u := c.BuildRequestUrl(path, nil)
-	var h http.Header
-	if header != nil {
-		h = header
-	} else {
-		h = http.Header{}
-	}
-	if body.Len() > 0 {
-		h.Set("Content-Type", "application/x-www-form-urlencoded")
-		h.Set("Content-Length", fmt.Sprintf("%d", body.Len()))
-	}
-
-	return c.sendHttp(verb, u.String(), h, body)
+	return c.sendFormString(verb, path, data, header)
 }
 
 func (c *Client) sendJsonRequest(verb, path string, params map[string]interface{}, header http.Header) ([]byte, error) {
@@ -150,6 +172,23 @@ func (c *Client) sendJsonRequest(verb, path string, params map[string]interface{
 	}
 	if body.Len() > 0 {
 		h.Set("Content-Type", "application/json")
+		h.Set("Content-Length", fmt.Sprintf("%d", body.Len()))
+	}
+
+	return c.sendHttp(verb, u.String(), h, body)
+}
+
+func (c *Client) sendFormString(verb, path string, data string, header http.Header) ([]byte, error) {
+	body := strings.NewReader(data)
+	u := c.BuildRequestUrl(path, nil)
+	var h http.Header
+	if header != nil {
+		h = header
+	} else {
+		h = http.Header{}
+	}
+	if body.Len() > 0 {
+		h.Set("Content-Type", "application/x-www-form-urlencoded")
 		h.Set("Content-Length", fmt.Sprintf("%d", body.Len()))
 	}
 
